Xcode 11: unset CODE_SIGN_IDENTITY for iOS SDK 13 so xcodebuild works (#75)
in Xcode 11, when CODE_SIGN_IDENTITY is left unchanged, the default
is set to "Apple Developer" (as seen in the settings UI), and this
causes attempts to build for Distribution to break for .app products
that include ZipUtilities.framework .

setting the value to "" (i.e. empty) eliminates the diagnostics and
allows the xcodebuild to proceed.